inventory-system/app/requirements.txt
mysql-connector-python
tabulate
fastapi
uvicorn
pydantic